Split autodeploy.py into multiple modules.
