+++ project/thscoreboard/thscoreboard/deploy/__init__.py
@@ -1,0 +1,1 @@
+"""Modules related to deploying new releases."""
