+++ project/thscoreboard/thscoreboard/urls.py
@@ -21,7 +21,7 @@ from django.urls import include, re_path, path
 from django.views.generic import base as base_views
 from replays.views import index
 from thscoreboard import settings
-from . import autodeploy
+from thscoreboard.deploy import views as deploy_views
 
 urlpatterns = [
     path('users/', include('users.urls')),
@@ -29,7 +29,7 @@ urlpatterns = [
     path('replays/', include('replays.urls.replay_urls')),
     path('docs/', include('replays.urls.docs_urls')),
     path('admin/', admin.site.urls),
-    path('deploy', autodeploy.deploy),
+    path('deploy', deploy_views.deploy),
 
     # When fetching non-HTML pages, the browser looks at /favicon.ico
     # to find an icon.
